upmiao/requirements.txt
asn1crypto==0.24.0
bootstrap-admin==0.4.0
cffi==1.11.5
cryptography==2.3.1
diff-match-patch==20121119
Django==2.1.2
django-appconf==1.0.2
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-import-export==1.1.0
django-reversion==3.0.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
jdcal==1.4
jieba==0.39
odfpy==1.3.6
openpyxl==2.5.8
pilkit==2.0
Pillow==5.2.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
Whoosh==2.7.4
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
